Created Project file for integration with eclipse
In order to have ease in editing and creating files to be uploaded to gihub repository, I transferred all the codes in  project folder so that we can easily edit this in Eclipse.

P.S. I transferred the data structure classes to a 'dataStructure' package for proper encapsulation of said classes.