Completed Needed Data Structure set with Runtime Exceptions
The Stack and Queue has been developed with minor testing, but it is believed to be robust. Also, the RuntimeException package was developed for "OutOfBounds" exception handling of data structures similar to the ArrayIndexOutOfBounds exception of conventional arrays. Changes has been made to the older data structures to enable them to be generic, and many more. PseudoArray class was added with length() method to have access with the length of the array by counting the number of traversals between elements of its home data structure, the LinkedList.